+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/EventNumberLinker.java
@@ -2,7 +2,9 @@ package uk.gov.justice.services.eventsourcing.eventpublishing;
 
 import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
 
+import uk.gov.justice.services.eventsourcing.eventpublishing.configuration.EventLinkingWorkerConfig;
 import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.AdvisoryLockDataAccess;
+import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.CompatibilityModePublishedEventRepository;
 import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.LinkEventsInEventLogDatabaseAccess;
 
 import java.util.Optional;
@@ -21,6 +23,12 @@ public class EventNumberLinker {
     @Inject
     private AdvisoryLockDataAccess advisoryLockDataAccess;
 
+    @Inject
+    private CompatibilityModePublishedEventRepository compatibilityModePublishedEventRepository;
+
+    @Inject
+    private EventLinkingWorkerConfig eventLinkingWorkerConfig;
+
     @Transactional(REQUIRES_NEW)
     public boolean findAndAndLinkNextUnlinkedEvent() {
 
@@ -34,6 +42,13 @@ public class EventNumberLinker {
 
                 linkEventsInEventLogDatabaseAccess.linkEvent(eventId, newEventNumber, previousEventNumber);
                 linkEventsInEventLogDatabaseAccess.insertLinkedEventIntoPublishQueue(eventId);
+
+                // Temporary. To be removed once the migration to the new publishing is released
+                // and published_event table is deleted
+                if(eventLinkingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable()) {
+                    compatibilityModePublishedEventRepository.insertIntoPublishedEvent(eventId, newEventNumber, previousEventNumber);
+                    compatibilityModePublishedEventRepository.setEventNumberSequenceTo(newEventNumber);
+                }
 
                 return true;
             }