+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/LinkedEventPublisher.java
@@ -3,13 +3,10 @@ package uk.gov.justice.services.eventsourcing.eventpublishing;
 import static java.lang.String.format;
 import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
 
-import uk.gov.justice.services.eventsourcing.eventpublishing.configuration.EventPublishingWorkerConfig;
 import uk.gov.justice.services.eventsourcing.publishedevent.EventPublishingException;
-import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.CompatibilityModePublishedEventRepository;
 import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.EventPublishingRepository;
 import uk.gov.justice.services.eventsourcing.publisher.jms.EventPublisher;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.LinkedEvent;
-import uk.gov.justice.services.eventsourcing.repository.jdbc.event.MissingEventNumberException;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import java.util.Optional;
@@ -29,12 +26,6 @@ public class LinkedEventPublisher {
     @Inject
     private LinkedJsonEnvelopeCreator linkedJsonEnvelopeCreator;
 
-    @Inject
-    private EventPublishingWorkerConfig eventPublishingWorkerConfig;
-
-    @Inject
-    private CompatibilityModePublishedEventRepository compatibilityModePublishedEventRepository;
-
     @Transactional(REQUIRES_NEW)
     public boolean publishNextNewEvent() {
 
@@ -48,16 +39,7 @@ public class LinkedEventPublisher {
                 eventPublisher.publish(linkedJsonEnvelope);
                 eventPublishingRepository.setIsPublishedFlag(eventId.get(), true);
 
-                // Temporary. To be removed once the migration to the new publishing is released
-                // and published_event table is deleted
-                if(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable()) {
-                    compatibilityModePublishedEventRepository.insertIntoPublishedEvent(linkedJsonEnvelope);
-                    final Long eventNumber = linkedEvent
-                            .getEventNumber()
-                            .orElseThrow(() -> new MissingEventNumberException(format("Event with id '%s' has null event_number in event_log table", eventId.get())));
-                    compatibilityModePublishedEventRepository.setEventNumberSequenceTo(eventNumber);
-                }
-                
+
                 return true;
 
             } else {