+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/eventpublishing/configuration/EventLinkingWorkerConfigTest.java
@@ -44,4 +44,14 @@ public class EventLinkingWorkerConfigTest {
 
         assertThat(eventLinkingWorkerConfig.getTimeBetweenRunsMilliseconds(), is(milliseconds));
     }
+
+    @Test
+    public void shouldGetPublishEventToPublishedEventTable() throws Exception {
+
+        assertThat(eventLinkingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable(), is(false));
+
+        setField(eventLinkingWorkerConfig, "insertEventIntoPublishedEventTable", "true");
+
+        assertThat(eventLinkingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable(), is(true));
+    }
 }