+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/eventpublishing/configuration/EventPublishingWorkerConfigTest.java
@@ -45,13 +45,4 @@ public class EventPublishingWorkerConfigTest {
         assertThat(eventPublishingWorkerConfig.getTimeBetweenRunsMilliseconds(), is(milliseconds));
     }
 
-    @Test
-    public void shouldGetPublishEventToPublishedEventTable() throws Exception {
-
-        assertThat(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable(), is(false));
-
-        setField(eventPublishingWorkerConfig, "insertEventIntoPublishedEventTable", "true");
-
-        assertThat(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable(), is(true));
-    }
 }